requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
olefile==0.45.1
Pillow==4.1.1
pkg-resources==0.0.0
python-resize-image==1.1.11
requests==2.18.4
urllib3==1.22
